Updating package.json
Otherwise gocd-cypress/npm test on windows fails with:
Invalid tag name "$(npm": Tags may not have any characters that encodeURIComponent encodes.

I did not yet find a reliable way to get the file name dynamically.
Hardcoding the file name would be
better to avoid. If the version is not updated testing fail,
so at least we cannot forget to update it